edumaps_/backend/requirements.txt
bleach==6.2.0
boto3==1.39.15
Flask==3.1.1
flask_cors==6.0.1
flask_sqlalchemy==3.1.1
pandas==2.3.1
Pillow==11.3.0
python-magic-bin==0.4.14
redis==6.2.0
scikit_learn==1.7.1
SQLAlchemy==2.0.41
Werkzeug==3.1.3
